HH-218702 fix spring configuration
