+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/NabHibernateCommonConfig.java
@@ -1,24 +1,8 @@
 package ru.hh.nab.hibernate;
 
-import jakarta.annotation.Nullable;
-import jakarta.persistence.EntityManagerFactory;
-import java.util.function.Function;
-import static java.util.stream.Collectors.toMap;
-import java.util.stream.Stream;
-import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.annotation.Primary;
-import org.springframework.orm.jpa.JpaTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
-import ru.hh.nab.datasource.aspect.ExecuteOnDataSourceAspect;
-import ru.hh.nab.datasource.aspect.ExecuteOnDataSourceTransactionCallbackFactory;
-import ru.hh.nab.datasource.routing.RoutingDataSourceFactory;
-import ru.hh.nab.datasource.transaction.DataSourceContextTransactionManager;
-import ru.hh.nab.datasource.transaction.TransactionalScope;
-import ru.hh.nab.datasource.validation.DataSourcesReadyTarget;
-import ru.hh.nab.datasource.validation.ExecuteOnDataSourceBeanPostProcessor;
 import ru.hh.nab.hibernate.adapter.NabHibernateJpaVendorAdapter;
 import ru.hh.nab.hibernate.adapter.NabHibernatePersistenceProvider;
 import ru.hh.nab.hibernate.events.EventListenerRegistryPropagator;
@@ -28,35 +12,13 @@ import ru.hh.nab.jpa.NabJpaCommonConfig;
 
 @Configuration
 @Import({
-    RoutingDataSourceFactory.class,
     NabJpaCommonConfig.class,
     NabHibernateJpaVendorAdapter.class,
     NabHibernatePersistenceProvider.class,
     NabServiceContributor.class,
     EventListenerRegistryPropagator.class,
-    ExecuteOnDataSourceBeanPostProcessor.class,
-    DataSourcesReadyTarget.class,
-    TransactionalScope.class,
 })
 public class NabHibernateCommonConfig {
-
-  @Primary
-  @Bean
-  DataSourceContextTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
-    JpaTransactionManager jpaTransactionManager = new JpaTransactionManager(entityManagerFactory);
-    return new DataSourceContextTransactionManager(jpaTransactionManager);
-  }
-
-  @Bean
-  ExecuteOnDataSourceAspect executeOnDataSourceAspect(
-      ApplicationContext applicationContext,
-      @Nullable ExecuteOnDataSourceTransactionCallbackFactory transactionCallbackFactory
-  ) {
-    var txManagers = Stream
-        .of(applicationContext.getBeanNamesForType(PlatformTransactionManager.class))
-        .collect(toMap(Function.identity(), beanName -> applicationContext.getBean(beanName, PlatformTransactionManager.class)));
-    return new ExecuteOnDataSourceAspect(applicationContext.getBean(PlatformTransactionManager.class), txManagers, transactionCallbackFactory);
-  }
 
   @Bean
   ServiceSupplier<?> nabSessionFactoryBuilderServiceSupplier() {
