+++ nab-hibernate/src/main/java/ru/hh/nab/hibernate/NabHibernateProdConfig.java
@@ -9,15 +9,15 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import ru.hh.nab.common.properties.PropertiesUtils;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.datasource.NabDataSourceProdConfig;
 import ru.hh.nab.hibernate.monitoring.HibernateStatisticsSender;
 import ru.hh.nab.hibernate.properties.HibernatePropertiesProvider;
+import ru.hh.nab.jpa.NabJpaProdConfig;
 import ru.hh.nab.metrics.StatsDSender;
 
 @Configuration
 @Import({
     NabHibernateCommonConfig.class,
-    NabDataSourceProdConfig.class
+    NabJpaProdConfig.class
 })
 public class NabHibernateProdConfig {
