+++ nab-jpa/src/main/java/ru/hh/nab/jpa/NabJpaCommonConfig.java
@@ -10,17 +10,22 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Primary;
+import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.JpaVendorAdapter;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.orm.jpa.SharedEntityManagerCreator;
 import org.springframework.orm.jpa.persistenceunit.PersistenceManagedTypes;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
+import ru.hh.nab.datasource.NabDataSourceCommonConfig;
+import ru.hh.nab.datasource.transaction.DataSourceContextTransactionManager;
 import ru.hh.nab.jpa.aspect.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
 
 @Configuration
 @EnableTransactionManagement(order = 0)
 @EnableAspectJAutoProxy
 @Import({
+    NabDataSourceCommonConfig.class,
     ExecuteOnDataSourceTransactionCallbackFactoryImpl.class,
 })
 public class NabJpaCommonConfig {
@@ -57,5 +62,12 @@ public class NabJpaCommonConfig {
   @Bean
   public static EntityManager sharedEntityManager(EntityManagerFactory emf) {
     return SharedEntityManagerCreator.createSharedEntityManager(emf);
+  }
+
+  @Primary
+  @Bean
+  DataSourceContextTransactionManager transactionManager(EntityManagerFactory entityManagerFactory) {
+    JpaTransactionManager jpaTransactionManager = new JpaTransactionManager(entityManagerFactory);
+    return new DataSourceContextTransactionManager(jpaTransactionManager);
   }
 }
