+++ nab-jpa/src/main/java/ru/hh/nab/jpa/NabJpaProdConfig.java
@@ -1,0 +1,13 @@
+package ru.hh.nab.jpa;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import ru.hh.nab.datasource.NabDataSourceProdConfig;
+
+@Configuration
+@Import({
+    NabJpaCommonConfig.class,
+    NabDataSourceProdConfig.class,
+})
+public class NabJpaProdConfig {
+}
